backend/app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
etelemetry==0.3.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cache2==0.2.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
langsmith==0.1.143
llama-cloud==0.1.5
llama-index==0.11.23
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.23
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.5.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.3.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.14
llama_cpp_python==0.3.2
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
numpy==1.24.4
ollama==0.3.3
openai==0.28.0
orjson==3.10.11
packaging==24.2
pandas==2.0.3
pathlib==1.0.1
pendulum==3.0.0
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
prov==2.0.1
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.1.4
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scipy==1.10.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.38.6
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
time-machine==2.15.0
together==1.3.4
tqdm==4.67.0
traits==6.3.2
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.30.6
virtualenv==20.26.6
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
